오브젝트 먹었을 때 점수 증가

+++ Assets/02.Scripts/01.Manager/GameManager.cs
@@ -63,7 +63,7 @@ public class GameManager : MonoBehaviour
         scoreText.text = score.ToString("F0");
     }
 
-    public void IncreaseScore(int plusScore)
+    public void IncreaseScore(float plusScore)
     {
         score += plusScore;
     }
